[refactor] 패키지 이동 및 재구조화
