+++ src/main/java/store/greeting/config/AuditorAwareImpl.java
@@ -5,7 +5,6 @@ import org.springframework.data.domain.AuditorAware;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import store.greeting.plus.AuthTokenParser;
 
 public class AuditorAwareImpl implements AuditorAware<String > {
   @Override
