+++ src/main/java/store/greeting/mail/MailConfig.java
@@ -1,4 +1,4 @@
-package store.greeting.plus;
+package store.greeting.mail;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
