+++ src/main/java/store/greeting/social/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package store.greeting.plus;
+package store.greeting.social;
 
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
