+++ src/main/java/store/greeting/social/OAuthAttributes.java
@@ -1,4 +1,4 @@
-package store.greeting.plus;
+package store.greeting.social;
 
 import lombok.Builder;
 import lombok.Getter;
